+++ lib/aliyun_oss/bucket/policy.ex
@@ -1,0 +1,95 @@
+defmodule Aliyun.Oss.Bucket.Policy do
+  @moduledoc """
+  Bucket Policy 相关操作
+  """
+
+  import Aliyun.Oss.Bucket, only: [get_bucket: 3, put_bucket: 4, delete_bucket: 2]
+
+  alias Aliyun.Oss.Client.{Response, Error}
+
+  @type error() :: %Error{body: String.t(), status_code: integer(), parsed_details: map()} | atom()
+
+  @doc """
+  PutBucketPolicy接口用于为指定的存储空间（Bucket）设置授权策略（Policy)。
+
+  ## Examples
+
+      iex> policy = %{
+        "Statement" => [
+          %{
+            "Action" => ["oss:PutObject", "oss:GetObject"],
+            "Effect" => "Deny",
+            "Principal" => ["1234567890"],
+            "Resource" => ["acs:oss:*:1234567890:*/*"]
+          }
+        ],
+        "Version" => "1"
+      }
+      iex> Aliyun.Oss.Bucket.Policy.put("some-bucket", policy)
+      {:ok, %Aliyun.Oss.Client.Response{
+        data: "",
+        headers: [
+          {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          ...
+        ]
+      }}
+  """
+  @spec put(String.t(), map()) :: {:error, error()} | {:ok, Response.t()}
+  def put(bucket, %{} = policy) do
+    put_bucket(bucket, %{}, %{"policy" => nil}, Jason.encode!(policy))
+  end
+
+
+  @doc """
+  GetBucketPolicy用于获取指定存储空间（Bucket）的权限策略（Policy）。
+
+  ## Examples
+
+      iex> Aliyun.Oss.Bucket.Policy.get("some-bucket")
+      {:ok, %Aliyun.Oss.Client.Response{
+        data: %{
+          "Statement" => [
+            %{
+              "Action" => ["oss:PutObject", "oss:GetObject"],
+              "Effect" => "Deny",
+              "Principal" => ["1234567890"],
+              "Resource" => ["acs:oss:*:1234567890:*/*"]
+            }
+          ],
+          "Version" => "1"
+        },
+        headers: [
+          {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"},
+          ...
+        ]
+      }}
+  """
+  @spec get(String.t()) :: {:error, error()} | {:ok, Response.t()}
+  def get(bucket) do
+    get_bucket(bucket, %{}, %{"policy" => nil})
+  end
+
+  @doc """
+  DeleteBucketPolicy用于删除指定存储空间（Bucket）的权限策略（Policy）。
+
+  ## Examples
+
+      iex> Aliyun.Oss.Bucket.Policy.delete("some-bucket")
+      {:ok,
+      %Aliyun.Oss.Client.Response{
+        data: "",
+        headers: [
+          {"Server", "AliyunOSS"},
+          {"Date", "Fri, 11 Jan 2019 05:19:45 GMT"},
+          {"Content-Length", "0"},
+          {"Connection", "keep-alive"},
+          {"x-oss-request-id", "5C3000000000000000000000"},
+          {"x-oss-server-time", "90"}
+        ]
+      }}
+  """
+  @spec delete(String.t()) :: {:error, error()} | {:ok, Aliyun.Oss.Client.Response.t()}
+  def delete(bucket) do
+    delete_bucket(bucket, %{"policy" => nil})
+  end
+end